refactored all classes into one file
To use fileprivate methods on the backendAPIService (these methods should not be exposed in subclasses).
It is possible to delete classes and move everything into the backendapiservice-class, but is it really better? There is the possibility to use MARK-ers for each section, but this will work too, without changing other code from the project.